Move avatar overlay inside card for better layout
