+++ osu.Game/Screens/OnlinePlay/Matchmaking/Match/BeatmapSelect/BeatmapCardMatchmaking.cs
@@ -3,6 +3,8 @@
 
 using System.Linq;
 using osu.Framework.Allocation;
+using osu.Framework.Audio;
+using osu.Framework.Audio.Sample;
 using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Extensions.LocalisationExtensions;
 using osu.Framework.Graphics;
@@ -42,6 +44,8 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
 
         private FillFlowContainer idleBottomContent = null!;
         private BeatmapCardDownloadProgressBar downloadProgressBar = null!;
+
+        public AvatarOverlay SelectionOverlay = null!;
 
         [Resolved]
         private OverlayColourProvider colourProvider { get; set; } = null!;
@@ -240,6 +244,11 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
                                             Progress = { BindTarget = DownloadTracker.Progress }
                                         }
                                     }
+                                },
+                                SelectionOverlay = new AvatarOverlay
+                                {
+                                    Anchor = Anchor.TopRight,
+                                    Origin = Anchor.TopRight,
                                 }
                             }
                         }
@@ -314,6 +323,131 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.Match.BeatmapSelect
                 }
 
                 return items.ToArray();
+            }
+        }
+
+        public partial class AvatarOverlay : CompositeDrawable
+        {
+            private readonly Container<SelectionAvatar> avatars;
+
+            private Sample? userAddedSample;
+            private double? lastSamplePlayback;
+
+            public AvatarOverlay()
+            {
+                AutoSizeAxes = Axes.Both;
+
+                InternalChild = avatars = new Container<SelectionAvatar>
+                {
+                    AutoSizeAxes = Axes.X,
+                    Height = SelectionAvatar.AVATAR_SIZE,
+                };
+
+                Padding = new MarginPadding { Vertical = 5 };
+            }
+
+            [BackgroundDependencyLoader]
+            private void load(AudioManager audio)
+            {
+                userAddedSample = audio.Samples.Get(@"Multiplayer/player-ready");
+            }
+
+            public bool AddUser(APIUser user, bool isOwnUser)
+            {
+                if (avatars.Any(a => a.User.Id == user.Id))
+                    return false;
+
+                var avatar = new SelectionAvatar(user, isOwnUser);
+
+                avatars.Add(avatar);
+
+                if (lastSamplePlayback == null || Time.Current - lastSamplePlayback > OsuGameBase.SAMPLE_DEBOUNCE_TIME)
+                {
+                    userAddedSample?.Play();
+                    lastSamplePlayback = Time.Current;
+                }
+
+                updateAvatarLayout();
+
+                avatar.FinishTransforms();
+
+                return true;
+            }
+
+            public bool RemoveUser(int id)
+            {
+                if (avatars.SingleOrDefault(a => a.User.Id == id) is not SelectionAvatar avatar)
+                    return false;
+
+                avatar.PopOutAndExpire();
+                avatars.ChangeChildDepth(avatar, float.MaxValue);
+
+                updateAvatarLayout();
+
+                return true;
+            }
+
+            private void updateAvatarLayout()
+            {
+                const double stagger = 30;
+                const float spacing = 4;
+
+                double delay = 0;
+                float x = 0;
+
+                for (int i = avatars.Count - 1; i >= 0; i--)
+                {
+                    var avatar = avatars[i];
+
+                    if (avatar.Expired)
+                        continue;
+
+                    avatar.Delay(delay).MoveToX(x, 500, Easing.OutElasticQuarter);
+
+                    x -= avatar.LayoutSize.X + spacing;
+
+                    delay += stagger;
+                }
+            }
+
+            public partial class SelectionAvatar : CompositeDrawable
+            {
+                public const float AVATAR_SIZE = 30;
+
+                public APIUser User { get; }
+
+                public bool Expired { get; private set; }
+
+                private readonly MatchmakingAvatar avatar;
+
+                public SelectionAvatar(APIUser user, bool isOwnUser)
+                {
+                    User = user;
+                    Size = new Vector2(AVATAR_SIZE);
+
+                    InternalChild = avatar = new MatchmakingAvatar(user, isOwnUser)
+                    {
+                        Anchor = Anchor.Centre,
+                        Origin = Anchor.Centre,
+                    };
+                }
+
+                protected override void LoadComplete()
+                {
+                    base.LoadComplete();
+
+                    avatar.ScaleTo(0)
+                          .ScaleTo(1, 500, Easing.OutElasticHalf)
+                          .FadeIn(200);
+                }
+
+                public void PopOutAndExpire()
+                {
+                    avatar.ScaleTo(0, 400, Easing.OutExpo);
+
+                    this.FadeOut(100).Expire();
+                    Expired = true;
+                }
             }
         }
     }
